requirements.txt
# Core Backend
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Frontend
streamlit==1.28.0
streamlit-option-menu==0.3.6

# AI & ML
sentence-transformers>=2.5.1
faiss-cpu==1.7.4
google-generativeai>=0.3.2
transformers>=4.35.0
torch>=2.1.0

# Document Processing
PyMuPDF==1.23.8
python-docx==1.1.0
pdfplumber==0.10.0

# Data Processing
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2

# HTTP & API
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
aiofiles==23.2.1